feat: move summary interaction to client from server
