+++ app/(client)/home/@summary/page.tsx
@@ -1,13 +1,9 @@
 import ChatSummaryContainer from "@/components/client/chat-summary/container";
-import ComponentLoading from "@/components/utils/component-loading";
-import { Suspense } from "react";
 
-export default async function Summary({ searchParams }) {
+export default async function Summary() {
   return (
-    <Suspense fallback={<ComponentLoading text="Writing summary..." />}>
-      <div className="w-full">
-        <ChatSummaryContainer selectedAdvisor={searchParams?.advisorId} />
-      </div>
-    </Suspense>
+    <div className="w-full">
+      <ChatSummaryContainer />
+    </div>
   );
 }
