+++ components/client/chat-summary/summary-header.tsx
@@ -1,47 +1,30 @@
-import ComponentError from "@/components/utils/component-error";
-import { RedirectButton } from "@/components/utils/redirect-btn";
-import { createClient } from "@/lib/supabase/server";
-import Link from "next/link";
-import { redirect } from "next/navigation";
+"use client";
+import { Button } from "@/components/ui/button";
+import { Advisor } from "@/types/advisor";
 
-export default async function Header({ selectedAdvisor, user }) {
-  const supabase = createClient();
-
-  const { data: advisors, error: advisorsError } = await supabase
-    .from("matchings")
-    .select(
-      `
-        advisor_id,
-        advisor (
-          *
-        )
-      `
-    )
-    .eq("client_id", user.id);
-
-  if (!advisors) {
-    console.error("client/chat-summary/summary-header: fail to fetch advisors", advisorsError);
-    return <ComponentError message={"Fail to fetch advisors"} />;
-  }
-
-  if (!selectedAdvisor) {
-    redirect(`?advisorId=${advisors?.[0]?.advisor?.id}`);
-  }
-
+export default function Header({
+  advisors,
+  selected,
+  onSelect,
+}: {
+  advisors: Advisor[];
+  selected: string;
+  onSelect: (advisor: string) => void;
+}) {
   return (
     <div className="flex mb-6 border-b border-[#ECF0F3]">
-      {advisors.map(({ advisor }) => (
-        <Link
+      {advisors.map((advisor) => (
+        <Button
           key={advisor?.id}
-          href={`?advisorId=${advisor?.id}`}
-          className={`px-4 py-2 text-lg bg-transparent hover:bg-gray-200 border-0 outline-none shadow-none ring-0 ${
-            selectedAdvisor === advisor?.id
+          onClick={() => onSelect(advisor.id)}
+          className={`px-4 py-2 text-lg bg-transparent hover:bg-gray-200 border-0 outline-none shadow-none ring-0 rounded-none ${
+            selected === advisor?.id
               ? "text-[#5C59E4] border-b-2 border-[#5C59E4]"
               : "text-[#9CABC2]"
           }`}
         >
           {`${advisor?.first_name} ${advisor?.last_name}`}
-        </Link>
+        </Button>
       ))}
     </div>
   );
