+++ lib/actions/chat.ts
@@ -1,12 +1,14 @@
 "use server";
 
 import { Json } from "@/types/database.types";
+import { v4 as uuid } from "uuid";
 import {
   GENERATE_INITIAL_QUESTIONS_PROMPT,
   GENERATE_QUESTIONS_PROMPT,
+  SUMMARY_PROMPT,
 } from "../prompts";
 import { createClient } from "../supabase/server";
-import { callGPT4 } from "../utils";
+import { callGPT4, parseSummaryResponse } from "../utils/llm";
 import { getUserAnswerSummary } from "../utils/user-answer-summary";
 
 export const sendMessage = async (recipient: string, formData: FormData) => {
@@ -157,3 +159,91 @@ export const getSuggestions = async (recipient: string) => {
   }
   return (newSuggestion?.suggestions as string[]) || [];
 };
+
+export const getChatSummary = async (advisorId: string) => {
+  const supabase = createClient();
+  const { data: { user }, error: userError} = await supabase.auth.getUser();
+  if(!user){
+    console.error("client/chat-summary/summarizer: fail to fetch user", userError);
+    throw userError
+  }
+  const { data: summary, error: summaryError } = await supabase
+    .from("conversation_summaries")
+    .select()
+    .eq("advisor_id", advisorId)
+    .eq("client_id", user.id)
+    .limit(1)
+    .maybeSingle();
+
+  if (summaryError) {
+    console.error("client/chat-summary/summarizer: fail to fetch summary", summaryError);
+    throw summaryError;
+  }
+
+  const { data: messages, error: messagesError } = await supabase
+    .from("messages")
+    .select()
+    .or(
+      `and(sender.eq.${user.id},recipient.eq.${advisorId}),and(sender.eq.${advisorId},recipient.eq.${user.id})`
+    )
+    .order("created_at", { ascending: true });
+
+  if (!messages) {
+    console.error("client/chat-summary/summarizer: fail to fetch messages", messagesError);
+    throw messagesError;
+  }
+
+  if (
+    summary &&
+    summary.last_message_id === messages[messages.length - 1]?.id
+  ) {
+    return summary.summary ?? {}
+  }
+
+  const { data: client } = await supabase
+    .from("client")
+    .select()
+    .eq("id", user.id)
+    .single();
+
+  const previousSummary = `
+        Previous Summary:
+        Quick Summary: ${summary?.summary?.["quickSummary"]}
+        Main Points: ${(summary?.summary?.["mainPoints"] ?? []).join(", ")}
+        Services Offered: ${summary?.summary?.["servicesOffered"].join(", ")}
+        Analysis: ${summary?.summary?.["analysis"]}
+    `;
+  const conversation = messages
+    .map(
+      (msg) => `${msg.sender === user.id ? "User" : "Advisor"}: ${msg.message}`
+    )
+    .join("\n");
+
+  const prompt = SUMMARY_PROMPT.replace("{{CONVERSATION}}", conversation)
+    .replace(
+      "{{USER}}",
+      JSON.stringify(getUserAnswerSummary(client?.all_answers))
+    )
+    .replace("{{PREVIOUS_SUMMARY}}", previousSummary);
+
+  const response = await callGPT4(prompt, "");
+
+  const { data: newSummary, error } = await supabase
+    .from("conversation_summaries")
+    .upsert({
+      id: summary?.id || uuid(),
+      advisor_id: advisorId,
+      client_id: user.id,
+      last_message_id: messages[messages.length - 1]?.id,
+      summary: parseSummaryResponse(response),
+    })
+    .select()
+    .single();
+
+  if (error) {
+    console.error("client/chat-summary/summarizer: fail to upsert summary", error);
+    throw error
+  }
+
+  return newSummary.summary ?? {}
+}
